requirements.txt
# Core data science packages with versions
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.6.1
matplotlib>=3.7.0
seaborn==0.12.2
shap==0.42.1
imbalanced-learn==0.10.1

# Web and API dependencies
flask>=2.2.0
joblib>=1.2.0
pyjwt==2.8.0
nltk>=3.8.1
